Creation new function, big changes in mainCrentoid and other small changes.
In the Centroid function it is said to save the images with the centroids so that it does not create files continuously. The SearchOfLayers function is created to look for cases where it is outside the new polygon but inside layer 1. In the mainCentroid function there are still many changes.