Many changes
Add the Drawline3D function so that the layerVoronoi function
incorporates it into your code. What it does is to join the centroids of
the tracking of the same cell.
The LayerVoronoi (3D) changes are the incorporation of the function
previously said.

The old layerVoronoi function has been renamed to layerVoronoi2D.

Create new mainLayerNeighbours function, which calls
calculate_neigbours3D to calculate the neighbors of each cell. Then this
main function calculates how much surface it shares with each neighbor
and to which layer they belong.

Create a new function called transitions_layers that calculates the
transitions that each sample has and by layers.

Create a new 'test' function (the name will be changed). This function
redraws the images of those cells that only appear in a frame. If still
doing the treatment returns to appear in a single frame will eliminate
the cell and will be ordered the variable.

Limitation of other functions.