New functions, improvements, clean up and comments
The functions correctTracking, newTreatment and organizationTracking
correct the result of the sameCentroid function to make a more
exhaustive follow-up of the cells, since there were some cells that only
appeared in a frame. It has been improved the obtaining of transitions,
looking for Z that vary of number of neighbors with the function
transitions_layers.